@@ -34,7 +34,7 @@ void log() {
 
   loggerNoStack.w("Just a warning!");
 
-  logger.e("Error! Something bad happened", "Test Error");
+  logger.e("Error! Something bad happened", error: "Test Error");
 
   loggerNoStack.v({"key": 5, "value": "something"});
 
@@ -61,13 +61,9 @@ class HomeWidget extends StatelessWidget {
       body: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          LogConsoleOnShake(
-            dark: true,
-            child: Center(
-              child: Text("Shake Phone to open Console."),
-            ),
-          ),
-          FlatButton(onPressed: () => LogConsole.open(context), child: Text("or click here to open Console")),
+          TextButton(
+              onPressed: () => LogConsole.open(context),
+              child: Text("click here to open Console")),
         ],
       ),
     );
